+++ sdk/storage/azfile/CHANGELOG.md
@@ -15,7 +15,7 @@
 
 ### Other Changes
 
-* Updated version of azcore to 1.6.1 and azidentity to 1.3.0.
+* Updated version of azcore to 1.7.0 and azidentity to 1.3.0.
 
 ## 1.0.1 (Unreleased)